refactor: isolate session store test cleanup state
